Redefine MIPayload::NO_VALUE so that it hopefully compiles on mac

+++ src/snarl_distance_index.hpp
@@ -108,7 +108,7 @@ struct MIPayload {
     typedef std::uint64_t code_type; // We assume that this fits into gbwtgraph::payload_type.
 
     constexpr static code_type NO_CODE = std::numeric_limits<code_type>::max();
-    constexpr static std::uint64_t NO_VALUE = std::numeric_limits<size_t>::max(); 
+    constexpr static std::size_t NO_VALUE = std::numeric_limits<size_t>::max(); 
 
     const static size_t CHAIN_COMPONENT_OFFSET = 1;
     const static size_t CHAIN_COMPONENT_WIDTH = 8;
